Refactor: Refactor data generators
