+++ src/psycopmlutils/model_performance/__init__.py
@@ -1,3 +1,3 @@
 """Import model performance"""
 
-from .model_performance import *  # noqa
+from .model_performance import ModelPerformance  # noqa
